refactor: move distance utils to shared, fix dark mode in mobile styles
- Move distance.ts from MobileTasksView/utils/ to src/utils/ (used cross-module by WorkPage)
- Update MobileTasksView barrel to re-export from new location
- Update useWorkPage.ts import path
- Remove dead .mobile-tasks-container CSS rule from styles.ts
- Replace hardcoded #f3f4f6 background with dark-mode-aware equivalent
