+++ apps/web/src/components/MobileTasksView/styles.ts
@@ -28,17 +28,6 @@ export const mobileTasksStyles = `
   }
   .hide-scrollbar::-webkit-scrollbar { display: none; }
   .hide-scrollbar { -ms-overflow-style: none; scrollbar-width: none; }
-  .mobile-tasks-container {
-    position: fixed;
-    top: 0;
-    left: 0;
-    right: 0;
-    bottom: 0;
-    display: flex;
-    flex-direction: column;
-    background: #f3f4f6;
-    z-index: 1;
-  }
   .selected-marker {
     z-index: 1000 !important;
   }
