+++ apps/web/src/components/MobileTasksView/utils/distance.ts
@@ -1,38 +1,2 @@
-import { EARTH_RADIUS_KM } from '../../../constants/locations';
-
-/**
- * Haversine formula to calculate distance between two coordinates
- * @param lat1 - Latitude of first point
- * @param lon1 - Longitude of first point
- * @param lat2 - Latitude of second point
- * @param lon2 - Longitude of second point
- * @returns Distance in kilometers
- */
-export const calculateDistance = (
-  lat1: number,
-  lon1: number,
-  lat2: number,
-  lon2: number
-): number => {
-  const dLat = ((lat2 - lat1) * Math.PI) / 180;
-  const dLon = ((lon2 - lon1) * Math.PI) / 180;
-  const a =
-    Math.sin(dLat / 2) * Math.sin(dLat / 2) +
-    Math.cos((lat1 * Math.PI) / 180) *
-      Math.cos((lat2 * Math.PI) / 180) *
-      Math.sin(dLon / 2) *
-      Math.sin(dLon / 2);
-  const c = 2 * Math.atan2(Math.sqrt(a), Math.sqrt(1 - a));
-  return EARTH_RADIUS_KM * c;
-};
-
-/**
- * Format distance for display
- * @param km - Distance in kilometers
- * @returns Formatted string (e.g., "500m", "2.5km", "15km")
- */
-export const formatDistance = (km: number): string => {
-  if (km < 1) return `${Math.round(km * 1000)}m`;
-  if (km < 10) return `${km.toFixed(1)}km`;
-  return `${Math.round(km)}km`;
-};
+// Re-export from shared utils — distance is used cross-module (WorkPage, MobileTasksView)
+export { calculateDistance, formatDistance } from '../../../utils/distance';
